+++ packages/data/src/entities/action.ts
@@ -1,5 +1,5 @@
-import { IAction, IActionMap, Partial } from "ht-models";
-import * as _ from "underscore";
+import { IAction, IActionMap } from "ht-models";
+import _ from "underscore";
 import { HtPosition, IActionPositions } from "../interfaces";
 import moment from "moment-mini";
 import { GetDateRangeQuery } from "ht-utility";