+++ packages/data/src/utils/date-range.ts
@@ -1,25 +1,45 @@
 import * as moment from "moment-mini";
 
+export const DateMapService = (() => {
+  let instance;
+  return {
+    getInstance() {
+      if(!instance) {
+        instance = {
+          today_end:   moment().endOf('day').toISOString(),
+          today_start: moment().startOf('day').toISOString(),
+          yesterday_start: moment().subtract(1, 'days').toISOString(),
+          yesterday_end: moment().subtract(1, 'days').endOf('day').toISOString(),
+          day_7_start: moment().subtract(6, 'days').toISOString(),
+          month_start: moment().startOf('month').toISOString(),
+          day_30_start: moment().subtract(29, 'days').toISOString(),
+        };
+      }
+      return instance;
+    }
+  }
+
+})();
 export const DateRangeMap = {
   "today": {
-    start: moment().startOf('day').toISOString(),
-    end: moment().endOf('day').toISOString()
+    start: DateMapService.getInstance().today_start,
+    end: DateMapService.getInstance().today_end
   },
   "yesterday": {
-    start: moment().subtract(1, 'days').toISOString(),
-    end: moment().subtract(1, 'days').endOf('day').toISOString()
+    start: DateMapService.getInstance().yesterday_start,
+    end: DateMapService.getInstance().yesterday_end,
   },
   "last_7_days": {
-    start: moment().subtract(6, 'days').toISOString(),
-    end: moment().endOf('day').toISOString()
+    start: DateMapService.getInstance().day_7_start,
+    end: DateMapService.getInstance().today_end
   },
   "this_month": {
-    start: moment().startOf('month').toISOString(),
-    end: moment().endOf('day').toISOString()
+    start: DateMapService.getInstance().month_start,
+    end: DateMapService.getInstance().today_end
   },
   "last_30_days": {
-    start: moment().subtract(29, 'days').toISOString(),
-    end: moment().endOf('day').toISOString()
+    start: DateMapService.getInstance().day_30_start,
+    end: DateMapService.getInstance().today_end
   }
 };
 
@@ -57,5 +77,8 @@ export const DateRangeLabelMap = [
 ];
 
 export const isSameDateRange = (range1, range2) => {
-  return (range1.start == range2.start && range1.end == range2.end)
+  function nearTime(t1, t2) {
+    return Math.abs(new Date(t1).getTime() - new Date(t2).getTime()) < 1000;
+  }
+  return (nearTime(range1.start, range2.start) && nearTime(range1.end, range2.end))
 };