+++ packages/maps/src/map-utils/google-map-utils.ts
@@ -2,7 +2,6 @@ import { HtMarker, MapUtils } from "./interfaces";
 import * as _ from "underscore";
 import { HtPosition } from "ht-data";
 import { ITimeAwarePoint } from "ht-models";
-import {Color} from "../../../utils";
 declare var MarkerClusterer: any;
 declare var RichMarker: any;
 export function ExtendBounds(