+++ packages/models/CHANGELOG.md
@@ -1,2 +1,3 @@
-# 0.0.13-beta
+# 0.0.13
  - Add `IActionStatusGraph`
+ - Add `IActionsSummary`